Resolve circular dependency loading in some circumstances
Dependency loading was added to tube purpose to ensure the correct
loading of sub classes when factories were linted first. This should
also improve our ability to use class reloading. However this also
introduced a circular dependency when the linters were not used.

This change pulls the QCbehaviour out into a separate file to
allow it to be loaded indepdendently.
